type mismatch

+++ cs336_systems/flash_attention/flash_attention_triton.py
@@ -245,7 +245,7 @@ def flash_bwd_kernel(
         Pij = tl.exp(Sij - tl.broadcast_to(Li[:, None], Sij.shape))
 
         # (Bk, Bq) float16 * (Bq, D) float16 = (Bk, D) float16
-        dVi = tl.dot(tl.trans(Pij).to(dOi.dtype), dOi)
+        dVi = tl.dot(tl.trans(Pij).to(tl.float16), dOi)
         tl.atomic_add(dV_block_ptr, dVi.to(dV_block_ptr.type.element_ty))
 
         # (Bq, D) float16 * (D, Bk) float16 = (Bq, Bk) float16
